requirements.txt
pillow
scipy
numpy
matplotlib
pyglet
pyopengl
console-menu
meshcat